# open-api/immich-openapi-specs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/admin-sign-up": {
      "post": {
        "operationId": "signUpAdmin",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAdminResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/change-password": {
      "post": {
        "operationId": "changePassword",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserAdminResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "login",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginCredentialDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "logout",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/auth/validateToken": {
      "post": {
        "operationId": "validateAccessToken",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateAccessTokenResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/libraries/{id}/validate": {
      "post": {
        "operationId": "validate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateLibraryDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateLibraryResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Libraries"
        ]
      }
    },
...
  },
  "info": {
    "title": "Immich",
    "description": "Immich API",
    "version": "1.105.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "Bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "immich_access_token"
      },
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "ChangePasswordDto": {
        "properties": {
          "newPassword": {
            "example": "password",
            "minLength": 8,
            "type": "string"
          },
          "password": {
            "example": "password",
            "type": "string"
          }
        },
        "required": [
          "newPassword",
          "password"
        ],
        "type": "object"
      },
...
      "LoginCredentialDto": {
        "properties": {
          "email": {
            "example": "testuser@email.com",
            "type": "string"
          },
          "password": {
            "example": "password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
      "LoginResponseDto": {
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "isAdmin": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "profileImagePath": {
            "type": "string"
          },
          "shouldChangePassword": {
            "type": "boolean"
          },
          "userEmail": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "isAdmin",
          "name",
          "profileImagePath",
          "shouldChangePassword",
          "userEmail",
          "userId"
        ],
        "type": "object"
      },
      "LogoutResponseDto": {
        "properties": {
          "redirectUri": {
            "type": "string"
          },
          "successful": {
            "type": "boolean"
          }
        },
        "required": [
          "redirectUri",
          "successful"
        ],
        "type": "object"
      },
...
      "SignUpDto": {
        "properties": {
          "email": {
            "example": "testuser@email.com",
            "type": "string"
          },
          "name": {
            "example": "Admin",
            "type": "string"
          },
          "password": {
            "example": "password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "name",
          "password"
        ],
        "type": "object"
      },
...
      "UserAdminResponseDto": {
        "properties": {
          "avatarColor": {
            "$ref": "#/components/schemas/UserAvatarColor"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "isAdmin": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "oauthId": {
            "type": "string"
          },
          "profileImagePath": {
            "type": "string"
          },
          "quotaSizeInBytes": {
            "format": "int64",
            "nullable": true,
            "type": "integer"
          },
          "quotaUsageInBytes": {
            "format": "int64",
            "nullable": true,
            "type": "integer"
          },
          "shouldChangePassword": {
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/UserStatus"
          },
          "storageLabel": {
            "nullable": true,
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "avatarColor",
          "createdAt",
          "deletedAt",
          "email",
          "id",
          "isAdmin",
          "name",
          "oauthId",
          "profileImagePath",
          "quotaSizeInBytes",
          "quotaUsageInBytes",
          "shouldChangePassword",
          "status",
          "storageLabel",
          "updatedAt"
        ],
        "type": "object"
      },
...
      "UserAvatarColor": {
        "enum": [
          "primary",
          "pink",
          "red",
          "yellow",
          "blue",
          "green",
          "purple",
          "orange",
          "gray",
          "amber"
        ],
        "type": "string"
      },
...
      "UserStatus": {
        "enum": [
          "active",
          "removing",
          "deleted"
        ],
        "type": "string"
      },
...
      "ValidateAccessTokenResponseDto": {
        "properties": {
          "authStatus": {
            "type": "boolean"
          }
        },
        "required": [
          "authStatus"
        ],
        "type": "object"
      },
      "ValidateLibraryDto": {
        "properties": {
          "exclusionPatterns": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "importPaths": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
      "ValidateLibraryImportPathResponseDto": {
        "properties": {
          "importPath": {
            "type": "string"
          },
          "isValid": {
            "default": false,
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "importPath",
          "isValid"
        ],
        "type": "object"
      },
      "ValidateLibraryResponseDto": {
        "properties": {
          "importPaths": {
            "items": {
              "$ref": "#/components/schemas/ValidateLibraryImportPathResponseDto"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
    }
  }
}